dink
Changed the spammer a little and adjusted some other stuff.

+++ examples/ESPNOW_Stress_Test/ESPNOW_Stress_Test.ino
@@ -12,7 +12,7 @@ void setup() {
   beginFDRS();
 }
 void loop() {
-  for (int i=0; i < 255; i++) {
+  for (uint8_t i=0; i < 255; i++) {
     loadFDRS(float(i), IT_T);
     sendFDRS();
     delay(15);
